requirements.txt
docxcompose
tkcalendar
python-docx
tk
xlwt
pandas